Rework search UX

+++ desktop/src/main/java/time_tracker/component/search/SearchRecordPane.java
@@ -1,20 +1,25 @@
 package time_tracker.component.search;
 
-import io.github.palexdev.materialfx.controls.MFXToggleButton;
+import javafx.beans.binding.Bindings;
 import javafx.beans.binding.LongBinding;
 import javafx.beans.binding.StringBinding;
 import javafx.collections.ListChangeListener;
 import javafx.collections.ObservableList;
 import javafx.collections.WeakListChangeListener;
 import javafx.fxml.FXML;
+import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.input.MouseEvent;
+import javafx.scene.layout.Pane;
 import javafx.scene.layout.VBox;
 import lombok.extern.java.Log;
 import time_tracker.Utils;
 import time_tracker.common.GlobalContext;
 import time_tracker.common.annotation.NonNull;
+import time_tracker.component.TimeTrackerTabPane;
+import time_tracker.component.common.Icon;
 import time_tracker.config.properties.StopwatchProperties;
+import time_tracker.model.StopWatchAppState;
 import time_tracker.model.StopwatchRecord;
 import time_tracker.model.StopwatchRecordMeasurement;
 import time_tracker.service.AppStateService;
@@ -22,13 +27,14 @@ import time_tracker.service.AppStateService;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.logging.Level;
-import java.util.stream.Collectors;
 
 import static javafx.beans.binding.Bindings.createStringBinding;
 import static time_tracker.component.Utils.load;
+import static time_tracker.component.common.Confirmation.requireConfirmation;
+import static time_tracker.component.common.IconButton.initIconButton;
 
 @Log
-public class SearchRecordVBox extends VBox {
+public class SearchRecordPane extends Pane {
 
     @FXML
     private Label recordIdLabel;
@@ -37,21 +43,26 @@ public class SearchRecordVBox extends VBox {
     @FXML
     private Label totalTimeLabel;
     @FXML
-    private Label deleteLabel;
+    private Label totalTimeIconLabel;
     @FXML
-    private Label amountOfMeasurements;
-    @FXML
-    private MFXToggleButton trackedToggle;
+    private Button goToDateIconLabel;
     @FXML
     private VBox measurementsWrapperVBox;
+    @FXML
+    private Button trackButton;
+    @FXML
+    private Button notTrackButton;
+    @FXML
+    private Button deleteButton;
 
     private final AppStateService appStateService;
     private final StopwatchRecord record;
     // reference has to be stored as WeakListChangeListener is used to register it
     private final ListChangeListener<StopwatchRecordMeasurement> measurementsChangesListener;
+    private final StopWatchAppState stopWatchAppState;
 
-    public SearchRecordVBox(@NonNull final StopwatchRecord record) {
-        load("/fxml/search/SearchRecordVBox.fxml", this);
+    public SearchRecordPane(@NonNull final StopwatchRecord record) {
+        load("/fxml/search/SearchRecordPane.fxml", this);
         this.record = record;
 
         LocalDate date = record.getDate();
@@ -59,10 +70,8 @@ public class SearchRecordVBox extends VBox {
         // TODO make it a property
         dateLabel.textProperty().set(formattedLocalDate);
 
-        trackedToggle.selectedProperty()
-                .bindBidirectional(record.getTrackedProperty());
-
         this.appStateService = GlobalContext.get(AppStateService.class);
+        stopWatchAppState = GlobalContext.get(StopWatchAppState.class);
 
         // TODO it seems that listener has to be removed or to be moved to another place
         record.getTrackedProperty()
@@ -76,6 +85,20 @@ public class SearchRecordVBox extends VBox {
 
         bindTotalTime();
         measurementsChangesListener = bindAmountOfMeasurements();
+
+        initIconButton(trackButton, 20, Icon.CHECK);
+        trackButton.visibleProperty().bind(Bindings.not(this.record.getTrackedProperty()));
+        trackButton.managedProperty().bind(Bindings.not(this.record.getTrackedProperty()));
+
+        initIconButton(goToDateIconLabel, 12, Icon.INPUT);
+
+        initIconButton(notTrackButton, 20, Icon.CHECK, List.of("icon-button", "icon-button-green"), List.of("button-icon-green"));
+        notTrackButton.visibleProperty().bind(this.record.getTrackedProperty());
+        notTrackButton.managedProperty().bind(this.record.getTrackedProperty());
+
+        initIconButton(totalTimeIconLabel, 20, Icon.STOPWATCH, List.of("icon-label-black"), List.of("label-icon-black"));
+
+        initIconButton(deleteButton, 15, Icon.DELETE);
     }
 
     @NonNull
@@ -84,25 +107,18 @@ public class SearchRecordVBox extends VBox {
         var measurementsChangesListener = (ListChangeListener<StopwatchRecordMeasurement>) c -> {
             log.fine(() -> "stopwatch records have been changed. Actual " + c);
             ObservableList<? extends StopwatchRecordMeasurement> measurements = c.getList();
-            setAmountOfMeasurements(measurements);
             setMeasurementsReport(measurements);
         };
 
         stopwatchRecords.addListener(new WeakListChangeListener<>(measurementsChangesListener));
-        setAmountOfMeasurements(stopwatchRecords);
         setMeasurementsReport(stopwatchRecords);
         return measurementsChangesListener;
-    }
-
-    private void setAmountOfMeasurements(@NonNull final List<? extends StopwatchRecordMeasurement> measurements) {
-        int amountOfMeasurements = measurements.size();
-        this.amountOfMeasurements.setText(Integer.toString(amountOfMeasurements));
     }
 
     private void setMeasurementsReport(@NonNull final List<? extends StopwatchRecordMeasurement> measurements) {
         List<SearchMeasurementVBox> measurementVBoxes = measurements.stream()
                 .map(SearchMeasurementVBox::new)
-                .collect(Collectors.toList());
+                .toList();
 
         measurementsWrapperVBox.getChildren().clear();
         measurementsWrapperVBox.getChildren().addAll(measurementVBoxes);
@@ -112,7 +128,13 @@ public class SearchRecordVBox extends VBox {
     protected void delete(@NonNull final MouseEvent event) {
         log.log(Level.FINE, "deleteButton is clicked");
 
-        appStateService.delete(record);
+        requireConfirmation().whenComplete((it, ex) -> {
+            if (it) {
+                appStateService.delete(record);
+            } else {
+                log.fine(() -> "'No' button is clicked");
+            }
+        });
         event.consume();
     }
 
@@ -130,4 +152,30 @@ public class SearchRecordVBox extends VBox {
         totalTimeLabel.textProperty()
                 .bind(longBinding);
     }
+
+    @FXML
+    protected void track() {
+        log.log(Level.FINE, "'Track' button is clicked");
+        record.setTracked(true);
+    }
+
+    @FXML
+    protected void notTrack() {
+        log.log(Level.FINE, "'Not track' button is clicked");
+        record.setTracked(false);
+    }
+
+    @FXML
+    protected void goToDate() {
+        log.log(Level.FINE, "'Go to date' button is clicked");
+        LocalDate date = record.getDate();
+
+        log.log(Level.FINE, () ->
+                String.format("Date = '%s', record name = '%s', record id = '%s'", date, record.getName(), record.getId()));
+
+        stopWatchAppState.setChosenDate(date);
+        stopWatchAppState.getChosenStopwatchRecord().set(record);
+        TimeTrackerTabPane.INSTANCE.openStopWatchTab();
+    }
+
 }
